+++ src/action/processor.js
@@ -8,37 +8,32 @@ var Rubix = require('../core/rubix.js'),
     calc = require('../utils/calc.js');
 
 module.exports = function (action, framestamp, frameDuration) {
-    var props = action.props(),
-        data = action.data(),
-        values = action.values,
-        rubix = Rubix[props.rubix],
+    var values = action.values,
+        rubix = Rubix[action.rubix],
         valueRubix = rubix,
-        hasChanged = false,
         defaultRoute = routes.getName(),
         i = 0,
-        order = props.order = props.order || [],
+        order = action.order = action.order || [],
         orderLength = order.length,
         key = '', value, output;
     
     // Update elapsed
     if (rubix.updateInput) {
-        rubix.updateInput(action, props, frameDuration);
+        rubix.updateInput(action, frameDuration);
     }
 
     // Fire onStart if first frame
     if (action.firstFrame) {
         routes.shard(function (route, output) {
             if (route.onStart) {
-                route.onStart(output, action, values, props, data);
+                route.onStart(output, action, values);
             }
-        }, props);
-        
-        action.firstFrame = false;
+        }, action);
     }
     
     // Update Input if available
-    if (props.input) {
-        action.output.input = props.input.onFrame(framestamp);
+    if (action.input) {
+        action.output.input = action.input.onFrame(framestamp);
     }
 
     // Update values
@@ -54,7 +49,7 @@ module.exports = function (action, framestamp, frameDuration) {
         }
 
         // Calculate new value
-        output = valueRubix.process(key, value, values, props, action, frameDuration);
+        output = valueRubix.process(key, value, values, action, frameDuration);
         
         // Limit if range set
         if (valueRubix.limit) {
@@ -79,7 +74,7 @@ module.exports = function (action, framestamp, frameDuration) {
         
         // Check if changed and update
         if (value.current != output) {
-            hasChanged = true;
+            action.hasChanged = true;
         }
 
         // Set current and add unit (if any) for output
@@ -93,29 +88,30 @@ module.exports = function (action, framestamp, frameDuration) {
     routes.shard(function (route, output) {
         // Fire onFrame every frame
         if (route.onFrame) {
-            route.onFrame(output, action, values, props, data);
+            route.onFrame(output, action, values);
         }
         
         // Fire onChanged if values have changed
-        if (hasChanged && route.onChange) {
-            route.onChange(output, action, values, props, data);
+        if (action.hasChanged && route.onChange || action.firstFrame) {
+            route.onChange(output, action, values);
         }
     }, action.output);
 
     // Fire onEnd if ended
-    if (rubix.hasEnded(action, hasChanged)) {
+    if (rubix.hasEnded(action, action.hasChanged)) {
         action.isActive(false);
         
         routes.shard(function (route, output) {
             if (route.onEnd) {
-                route.onEnd(output, action, values, props, data);
+                route.onEnd(output, action, values);
             }
         }, action.output);
         
-        if (!action.isActive() && props.rubix === 'play') {
+        if (!action.isActive() && action.rubix === 'play') {
             action.next();
         }
     }
-    
+
+    action.firstFrame = false;
     action.framestamp = framestamp;
 };