+++ src/routes/path.js
@@ -4,13 +4,13 @@ var createStyles = require('./path/builder.js');
 
 module.exports = {
 
-    onStart: function (bucket, action, values, props) {
-        if (props.dom) {
-            props.pathLength = props.dom.getTotalLength();
+    onStart: function (bucket, action) {
+        if (action.dom) {
+            action.pathLength = action.dom.getTotalLength();
         }
     },
     
-    onChange: function (output, action, values, props) {
-        action.style(createStyles(output, props.pathLength));
+    onChange: function (output, action) {
+        action.style(createStyles(output, action.pathLength));
     }
 };